+++ StockBox/changelog.md
@@ -1,4 +1,4 @@
-# Stockbox Changelog
+# StockBox Changelog
 
 All notable changes to this project will be documented in this file.